+++ Example/BaseClasses/NavigationController.swift
@@ -1,0 +1,13 @@
+//
+//  NavigationController.swift
+//  BaseClasses
+//
+//  Created by Anton Plebanovich on 8/7/17.
+//  Copyright © 2017 CocoaPods. All rights reserved.
+//
+
+import UIKit
+import BaseClasses
+
+
+class NavigationController: BaseClasses.NavigationController {}