+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 All notable changes to this project will be documented in this file.
 `BaseClasses` adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0](https://github.com/APUtils/BaseClasses/releases/tag/1.2.0)
+Released on 07/11/2017.
+
+#### Added
+- NavigationBar class to allow scroll touches
+  - Added by [Anton Plebanovich](https://github.com/anton-plebanovich).
+
+## [1.1.0](https://github.com/APUtils/BaseClasses/releases/tag/1.1.0)
+Released on 07/11/2017.
+
+#### Added
+- NavigationController to return status bar style from child controllers
+- FullSizeCollectionView that resizes cells to fit its size
+  - Added by [Anton Plebanovich](https://github.com/anton-plebanovich).
+
 ## [1.0.0](https://github.com/APUtils/BaseClasses/releases/tag/1.0.0)
 Released on 07/11/2017.